extra files

+++ nopen/test.c
@@ -11,30 +11,31 @@ int main(void){
     int fd;
     char c;
 
-    fd = open("test.c", O_RDONLY);
+    fd = open("rules.txt", O_RDONLY);
     fp = fdopen(fd, "r");
-    if (fp == NULL) 
-    {
+    if (fp == NULL) {
         printf("Cannot open file \n");
         return 1;
     }
     // Read contents from file
+    printf("\nReading file rules...\n");
     c = fgetc(fp);
     while (c != EOF) {
         printf ("%c", c);
         c = fgetc(fp);
     } 
-
+    // stat
     struct stat sb;
-
-    stat("test.c",&sb);
-    printf("%ld\n",sb.st_size);
-    
-    fd = open("hello.txt", O_RDWR | O_CREAT);
+    if (stat("test.c",&sb))
+        return 0;
+    printf("\nFile size: %ld \n", sb.st_size); 
+    // Creating a file
+    fd = open("hello.txt", O_RDWR | O_CREAT, S_IRWXU | S_IRWXG);
     fp = fdopen(fd, "w+");
-
     fprintf(fp,"hello world");
-
+    printf("Created file...\n");
+    //delete
     unlink("hello.txt");
+    printf("Deleted file...\n");
     return 0;
 }
